facial_recognition_system/requirements.txt
django
djangorestframework
opencv-python
numpy
face_recognition
